+++ src/frontend/js/data/SessionProvider/JoanieSessionProvider.spec.tsx
@@ -1,11 +1,14 @@
 import fetchMock from 'fetch-mock';
 import { IntlProvider } from 'react-intl';
-import { QueryClientProvider } from 'react-query';
+import { QueryClient, QueryClientProvider } from 'react-query';
 import { act, render, waitFor } from '@testing-library/react';
 import { ContextFactory as mockContextFactory, FonzieUserFactory } from 'utils/test/factories';
 import createQueryClient from 'utils/react-query/createQueryClient';
 import { Deferred } from 'utils/test/deferred';
-import { RICHIE_USER_TOKEN } from 'settings';
+import { REACT_QUERY_SETTINGS, RICHIE_USER_TOKEN } from 'settings';
+import { renderHook } from '@testing-library/react-hooks';
+import { useSession } from 'data/SessionProvider/index';
+import { PropsWithChildren } from 'react';
 import JoanieSessionProvider from './JoanieSessionProvider';
 
 jest.mock('utils/errors/handle');
@@ -16,30 +19,45 @@ jest.mock('utils/context', () => ({
     joanie_backend: { endpoint: 'https://joanie.endpoint.test' },
   }).generate(),
 }));
+jest.mock('utils/indirection/window', () => ({
+  location: {
+    assign: jest.fn(),
+  },
+}));
 
 // - Joanie Session Provider test suite
 describe('JoanieSessionProvider', () => {
+  const Wrapper = ({ client, children }: PropsWithChildren<{ client: QueryClient }>) => (
+    <IntlProvider locale="en">
+      <QueryClientProvider client={client}>
+        <JoanieSessionProvider>{children}</JoanieSessionProvider>
+      </QueryClientProvider>
+    </IntlProvider>
+  );
+
   beforeEach(() => {
     fetchMock.restore();
+    jest.useFakeTimers();
+    sessionStorage.clear();
+
+    fetchMock
+      .get('https://joanie.endpoint.test/api/addresses/', [])
+      .get('https://joanie.endpoint.test/api/credit-cards/', [])
+      .get('https://joanie.endpoint.test/api/orders/', []);
+  });
+
+  afterEach(() => {
+    jest.runOnlyPendingTimers();
+    jest.useRealTimers();
   });
 
   it('stores user access token within session storage', async () => {
     const queryClient = createQueryClient();
     const user = FonzieUserFactory.generate();
 
-    fetchMock
-      .get('https://auth.endpoint.test/api/v1.0/user/me', user)
-      .get('https://joanie.endpoint.test/api/addresses/', [])
-      .get('https://joanie.endpoint.test/api/credit-cards/', [])
-      .get('https://joanie.endpoint.test/api/orders/', []);
+    fetchMock.get('https://auth.endpoint.test/api/v1.0/user/me', user);
 
-    render(
-      <IntlProvider locale="en">
-        <QueryClientProvider client={queryClient}>
-          <JoanieSessionProvider />
-        </QueryClientProvider>
-      </IntlProvider>,
-    );
+    render(<Wrapper client={queryClient} />);
 
     await waitFor(() => {
       expect(fetchMock.lastUrl()).toEqual('https://auth.endpoint.test/api/v1.0/user/me');
@@ -53,19 +71,9 @@ describe('JoanieSessionProvider', () => {
     const user = FonzieUserFactory.generate();
     const deferredUser = new Deferred();
 
-    fetchMock
-      .get('https://auth.endpoint.test/api/v1.0/user/me', deferredUser.promise)
-      .get('https://joanie.endpoint.test/api/addresses/', [])
-      .get('https://joanie.endpoint.test/api/credit-cards/', [])
-      .get('https://joanie.endpoint.test/api/orders/', []);
+    fetchMock.get('https://auth.endpoint.test/api/v1.0/user/me', deferredUser.promise);
 
-    render(
-      <IntlProvider locale="en">
-        <QueryClientProvider client={queryClient}>
-          <JoanieSessionProvider />
-        </QueryClientProvider>
-      </IntlProvider>,
-    );
+    render(<Wrapper client={queryClient} />);
 
     await act(async () => deferredUser.resolve(user));
 
@@ -83,13 +91,7 @@ describe('JoanieSessionProvider', () => {
 
     fetchMock.get('https://auth.endpoint.test/api/v1.0/user/me', deferredUser.promise);
 
-    render(
-      <IntlProvider locale="en">
-        <QueryClientProvider client={queryClient}>
-          <JoanieSessionProvider />
-        </QueryClientProvider>
-      </IntlProvider>,
-    );
+    render(<Wrapper client={queryClient} />);
 
     await act(async () => deferredUser.resolve(null));
 
@@ -97,5 +99,73 @@ describe('JoanieSessionProvider', () => {
 
     expect(fetchMock.calls()).toHaveLength(1);
     expect(fetchMock.lastUrl()).toEqual('https://auth.endpoint.test/api/v1.0/user/me');
+  });
+
+  it('clears session storage on login', async () => {
+    const queryClient = createQueryClient({ persistor: true });
+    const user = FonzieUserFactory.generate();
+    const userDeferred = new Deferred();
+
+    fetchMock.get('https://auth.endpoint.test/api/v1.0/user/me', userDeferred.promise);
+    const { result, rerender } = renderHook(useSession, {
+      wrapper: Wrapper,
+      initialProps: { client: queryClient },
+    });
+
+    await act(async () => {
+      userDeferred.resolve(user);
+    });
+
+    await act(async () => {
+      jest.runOnlyPendingTimers();
+    });
+
+    expect(result.current.user).toStrictEqual(user);
+    expect(sessionStorage.getItem(REACT_QUERY_SETTINGS.cacheStorage.key)).toContain(user.username);
+    expect(sessionStorage.getItem(RICHIE_USER_TOKEN)).toEqual(user.access_token);
+
+    await act(async () => {
+      result.current.login();
+    });
+
+    expect(sessionStorage.getItem(REACT_QUERY_SETTINGS.cacheStorage.key)).toBeNull();
+    expect(sessionStorage.getItem(RICHIE_USER_TOKEN)).toBeNull();
+
+    rerender();
+    expect(result.current.user).toBeUndefined();
+  });
+
+  it('clears session storage on register', async () => {
+    const queryClient = createQueryClient({ persistor: true });
+    const user = FonzieUserFactory.generate();
+    const userDeferred = new Deferred();
+
+    fetchMock.get('https://auth.endpoint.test/api/v1.0/user/me', userDeferred.promise);
+    const { result, rerender } = renderHook(useSession, {
+      wrapper: Wrapper,
+      initialProps: { client: queryClient },
+    });
+
+    await act(async () => {
+      userDeferred.resolve(user);
+    });
+
+    await act(async () => {
+      jest.runOnlyPendingTimers();
+    });
+
+    expect(result.current.user).toStrictEqual(user);
+    expect(sessionStorage.getItem(REACT_QUERY_SETTINGS.cacheStorage.key)).toContain(user.username);
+    expect(sessionStorage.getItem(RICHIE_USER_TOKEN)).toEqual(user.access_token);
+
+    await act(async () => {
+      result.current.register();
+    });
+
+    expect(sessionStorage.getItem(REACT_QUERY_SETTINGS.cacheStorage.key)).toBeNull();
+    expect(sessionStorage.getItem(RICHIE_USER_TOKEN)).toBeNull();
+
+    rerender();
+    expect(result.current.user).toBeUndefined();
   });
 });